change Int to Long for identidiers

+++ app/com/superxtra/notepad/model/Note.scala
@@ -2,7 +2,7 @@ package com.superxtra.notepad.model
 
 import play.api.libs.json.{Json, OFormat, Reads}
 
-case class Note(id: Int, title: String, body: String, userId: Int)
+case class Note(id: Long, title: String, body: String, userId: Long)
 
 object Note {
   implicit val userJsonFormat: OFormat[Note] = Json.format[Note]
